ci(env): ci build

+++ packages/burnRegistry/contracts/Migrations.sol
@@ -2,7 +2,7 @@ pragma solidity 0.5.10;
 
 contract Migrations {
     address public owner;
-    uint public last_completed_migration;
+    uint256 public last_completed_migration;
 
     constructor() public {
         owner = msg.sender;
@@ -12,7 +12,7 @@ contract Migrations {
         if (msg.sender == owner) _;
     }
 
-    function setCompleted(uint completed) external restricted {
+    function setCompleted(uint256 completed) external restricted {
         last_completed_migration = completed;
     }
